requirements.txt
# Core AI & Transformers
transformers==4.41.0
tokenizers==0.19.1
huggingface_hub==0.30.2
datasets==2.19.1
peft==0.10.0
accelerate==0.30.1
sentence-transformers==4.1.0
triton==2.2.0
bitsandbytes>=0.43.1

# LangChain Stack
langchain==0.3.23
langchain-core==0.3.52
langchain-community==0.3.21
langchain-openai==0.3.12
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langchain-huggingface==0.1.2
langchain-chroma==0.2.2

# Vector DBs
chromadb==0.6.3
faiss-cpu==1.10.0

# File Parsing
pypdf==5.4.0
pymupdf==1.25.5
beautifulsoup4
bs4

# Search & Metadata
wikipedia
arxiv
unstructured
youtube_transcript_api
pytube
duckduckgo_search==5.3.1b1

# Database
mysql-connector-python==9.3.0
SQLAlchemy==2.0.40
duckdb==1.2.2

# Data
pandas==2.2.3
numexpr==2.10.2
numpy==1.26.4

# Streamlit & Web
streamlit
blinker
cachetools
validators==0.28.1
python-dotenv
ipykernel
protobuf==5.27.0
sympy==1.13.1
click
toml
altair
pydeck
watchdog
tornado
gitpython
